+++ snuba/datasets/events.py
@@ -23,8 +23,8 @@ from snuba.query.extensions import (
     PerformanceExtension,
     ProjectExtension,
     QueryExtension,
-    TimeSeriesExtension,
 )
+from snuba.query.timeseries import TimeSeriesExtension
 
 from snuba.util import (
     alias_expr,
